+++ main-site/lib/getStatic/celebTagPage.getStaticProps.ts
@@ -19,14 +19,11 @@ export type TagPageProps = NonNullable<
 export const getStaticProps = async ({
   params,
 }: {
-  params: { celeb: string; celebTagId: string };
+  params: { slug: string; tagId: string };
 }) => {
-  log(
-    'info',
-    `tagPage getStaticProps called: ${params.celeb}/${params.celebTagId}`,
-  );
+  log('info', `tagPage getStaticProps called: ${params.slug}/${params.tagId}`);
 
-  const results = await getCelebWithTimeline(params.celeb, true);
+  const results = await getCelebWithTimeline(params.slug, true);
 
   if (!results) {
     return {
@@ -34,21 +31,21 @@ export const getStaticProps = async ({
     };
   }
 
-  if (!tagExists(results.celeb.tagTimeline, params.celebTagId)) {
+  if (!tagExists(results.celeb.tagTimeline, params.tagId)) {
     return {
       notFound: true,
     };
   }
 
   const tagFacts = results.celeb.facts.filter((f) =>
-    f.tags.some((t) => t.tag._id === params.celebTagId),
+    f.tags.some((t) => t.tag._id === params.tagId),
   );
-  const tag = tagFacts[0].tags.find((t) => t.tag._id === params.celebTagId)!;
+  const tag = tagFacts[0].tags.find((t) => t.tag._id === params.tagId)!;
 
   const { otherCelebsWithTag, otherCelebsWithIssue } = await getRelatedCelebs(
-    params.celebTagId,
+    params.tagId,
     tag.tag.issue._id,
-    params.celeb,
+    params.slug,
     uniq([tag.tag.issue.name, ...results.orderOfIssues]),
   );
